Move xarray_image_as_png to data
This makes it possible to run
update_ranges without Pillow.

Refs #1549
