+++ tests/test_ogc_utils.py
@@ -12,6 +12,7 @@ import pytest
 import xarray
 from odc.geo.geom import polygon
 
+import datacube_ows.data
 import datacube_ows.http_utils
 import datacube_ows.ogc_utils
 import datacube_ows.time_utils
@@ -313,7 +314,7 @@ def test_png_loop_over() -> None:
             "alpha": dummy_da(200, "alpha", xyt_coords, dtype="uint8"),
         }
     )
-    imgs = datacube_ows.ogc_utils.xarray_image_as_png(data, loop_over="time")
+    imgs = datacube_ows.data.xarray_image_as_png(data, loop_over="time")
     assert len(imgs) == 2
     assert len(imgs[0]) == 78
     assert imgs[0].find(b"\x89PNG") == 0
@@ -328,9 +329,7 @@ def test_png_loop_over_anim() -> None:
             "alpha": dummy_da(200, "alpha", xyt_coords, dtype="uint8"),
         }
     )
-    imgs = datacube_ows.ogc_utils.xarray_image_as_png(
-        data, loop_over="time", animate=True
-    )
+    imgs = datacube_ows.data.xarray_image_as_png(data, loop_over="time", animate=True)
     assert len(imgs) == 173
     assert imgs.find(b"\x89PNG") == 0
 
@@ -344,7 +343,7 @@ def test_render_frame() -> None:
             "alpha": dummy_da(200, "alpha", xy_coords, dtype="uint8"),
         }
     )
-    png = datacube_ows.ogc_utils.render_frame(data, 5, 2)
+    png = datacube_ows.data.render_frame(data, 5, 2)
     assert png.shape == (2, 5, 4)
     data = xarray.Dataset(
         {
@@ -353,7 +352,7 @@ def test_render_frame() -> None:
             "blue": dummy_da(150, "blue", xy_coords, dtype="uint8"),
         }
     )
-    png = datacube_ows.ogc_utils.render_frame(data, 5, 2)
+    png = datacube_ows.data.render_frame(data, 5, 2)
     assert png.shape == (2, 5, 4)
